добавален модуль create-offers.js
