+++ js/data.js
@@ -1,5 +1,3 @@
-import {getRandomInteger, getRandomRealNumber, getRandomArrayElement, getRandomArrayPart, getNumberWithLeadZero} from './util.js';
-
 const OFFER_COUNT = 10;
 
 const TYPES = [
@@ -57,36 +55,4 @@ const PHOTOS_LINKS = [
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
 ];
 
-const createOffer = (i) => {
-  const lat = getRandomRealNumber(35.65, 35.7, 5);
-  const lng = getRandomRealNumber(139.7, 139.8, 5);
-  const checks = [getRandomInteger(0, TIMES.length - 1), getRandomInteger(0, TIMES.length - 1)];
-  return {
-    author: {
-      avatar: `img/avatars/user${getNumberWithLeadZero(i + 1)}.png`,
-    },
-    offer: {
-      title: TITLES[i],
-      address: `${lat}, ${lng}`,
-      price: getRandomInteger(1, 100),
-      type: getRandomArrayElement(TYPES),
-      rooms: getRandomInteger(1, 10),
-      guests: getRandomInteger(1, 10),
-      checkin: TIMES[Math.min(...checks)],
-      checkout: TIMES[Math.max(...checks)],
-      features: getRandomArrayElement(FEATURES),
-      description: DESCRIPTIONS[i],
-      photos: getRandomArrayPart(PHOTOS_LINKS),
-    },
-    location: {
-      lat,
-      lng,
-    }
-  };
-};
-
-function createOffers() {
-  return Array.from({ length: OFFER_COUNT }, (el, i) => createOffer(i));
-}
-
-export {createOffers};
+export {OFFER_COUNT, TYPES, TIMES, FEATURES, DESCRIPTIONS, TITLES, PHOTOS_LINKS};
